Update README.md
Might want to work on a better project description in the future(?)
Have a basic "Here's how you install and get started" done in the "Install" and "Documentation/Wiki" sections, will need to be expanded upon in the future I'm sure.
Credits are far from complete. Will need to find a way to link to and credit the artists and writing teams. For now, I've just linked to everyone(?) with access to the project on GitHub.
License is currently undetermined to my knowledge