Update README.md
Might want to work on a better project description in the future(?)
Have a basic "Here's how you install and get started" done in the "Install" and "Documentation/Wiki" sections, will need to be expanded upon in the future I'm sure.
Credits are far from complete. Will need to find a way to link to and credit the artists and writing teams. For now, I've just linked to everyone(?) with access to the project on GitHub. Linking either directly to Discord profiles or to various Linkedin/Newgrounds/Picarto.tv/etc. pages where the contributor could be contacted/their work viewed might be a good idea.
License is under consideration in the Discord ATM.